// src/app/shared/interfaces/result.ts
export interface Result {
  accType: string;
  network: string;
  companyImage: string;
  companyName: string;
  companyCode: string;
  labelStatus: string;
  status: string;
  companyNick: string;
  companyCountryVAT: string;
  companyGLN: string;
  companyAddress: string;
  companyCity: string;
  companyZIPCode: string;
  companyUrl: string;
  companyEmail:string;
  companyPhone: string;
}